server/requirements.txt
amqp==5.2.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
email-validator==2.1.1
flask==3.0.3
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.3
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
passlib==1.7.4
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
SQLAlchemy==2.0.29
typing-extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
werkzeug==3.0.2
wtforms==3.1.2
zipp==3.18.1
